Move internal/http/server
